backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
supabase==2.5.1
python-multipart
pyjwt
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.7.4
jinja2==3.1.4
pdfkit==1.0.0
